wh2026-backend/pyproject.toml
[project]
name = "wh2026-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "attrs>=25.4.0",
    "cattrs>=25.3.0",
    "python-socketio>=5.16.1",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.15.0",
]
